Rename Barista Actions to Interactions
As decided in https://github.com/SchibstedSpain/Barista/issues/132
we renamed our "action" classes as "interaction" to avoid confussion with
Espresso's view actions.

Barista Interactions are method's that perform an UI interaction upon calling.
On the other hand, Espresso ViewActions are objects that get executed by some other component.

In Barista we have both, since some of our Interactions are implemented with custom ViewActions.

Will be hard getting used to the new naming, but hopefuly will make our lives easier in the long term.
