+++ library/src/main/java/com/schibsted/spain/barista/interaction/BaristaEditTextInteractions.kt
@@ -1,13 +1,13 @@
-package com.schibsted.spain.barista.action
+package com.schibsted.spain.barista.interaction
 
 import android.support.annotation.IdRes
 import android.support.test.espresso.Espresso.onView
 import android.support.test.espresso.action.ViewActions.replaceText
 import android.support.test.espresso.matcher.ViewMatchers.withId
-import com.schibsted.spain.barista.action.BaristaScrollActions.safelyScrollTo
+import com.schibsted.spain.barista.interaction.BaristaScrollInteractions.safelyScrollTo
 import com.schibsted.spain.barista.internal.matcher.DisplayedMatchers.displayedAnd
 
-object BaristaEditTextActions {
+object BaristaEditTextInteractions {
 
     @JvmStatic
     fun writeTo(@IdRes editTextId: Int, text: String) {
