+++ library/src/main/java/com/schibsted/spain/barista/interaction/BaristaRadioButtonInteractions.kt
@@ -1,4 +1,4 @@
-package com.schibsted.spain.barista.action
+package com.schibsted.spain.barista.interaction
 
 import android.support.annotation.IdRes
 import android.support.test.espresso.Espresso.onView
@@ -7,10 +7,10 @@ import android.support.test.espresso.matcher.ViewMatchers.*
 import android.view.View
 import com.schibsted.spain.barista.internal.matcher.DisplayedMatchers.displayedAnd
 import com.schibsted.spain.barista.internal.matcher.HelperMatchers.atPosition
-import com.schibsted.spain.barista.action.BaristaScrollActions.safelyScrollTo
+import com.schibsted.spain.barista.interaction.BaristaScrollInteractions.safelyScrollTo
 import org.hamcrest.Matchers.allOf
 
-object BaristaRadioButtonActions {
+object BaristaRadioButtonInteractions {
 
     @JvmStatic
     fun clickRadioButtonItem(@IdRes radioGroupId: Int, @IdRes itemToClickId: Int) {
