+++ library/src/main/java/com/schibsted/spain/barista/interaction/BaristaSpinnerInteractions.java
@@ -1,4 +1,4 @@
-package com.schibsted.spain.barista.action;
+package com.schibsted.spain.barista.interaction;
 
 import android.support.annotation.IdRes;
 import android.support.test.espresso.action.ViewActions;
@@ -7,13 +7,13 @@ import android.widget.AdapterView;
 import static android.support.test.espresso.Espresso.onData;
 import static android.support.test.espresso.matcher.ViewMatchers.isAssignableFrom;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static com.schibsted.spain.barista.action.BaristaClickActions.clickOn;
+import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.anything;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 
-public class BaristaSpinnerActions {
+public class BaristaSpinnerInteractions {
 
   public static void clickSpinnerItem(@IdRes int id, int position) {
     clickOn(id);
