+++ library/src/main/java/com/schibsted/spain/barista/interaction/BaristaSwipeRefreshInteractions.kt
@@ -1,4 +1,4 @@
-package com.schibsted.spain.barista.action
+package com.schibsted.spain.barista.interaction
 
 import android.support.annotation.IdRes
 import android.support.v4.widget.SwipeRefreshLayout
@@ -8,7 +8,7 @@ import android.support.test.espresso.matcher.ViewMatchers.isAssignableFrom
 import android.support.test.espresso.matcher.ViewMatchers.withId
 import com.schibsted.spain.barista.internal.viewaction.SwipeRefreshActions.pullToRefresh
 
-object BaristaSwipeRefreshActions {
+object BaristaSwipeRefreshInteractions {
 
     @JvmStatic
     @JvmOverloads
