+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/AutoCompleteTextViewTest.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
-import static com.schibsted.spain.barista.action.BaristaAutoCompleteTextViewActions.writeToAutoComplete;
+import static com.schibsted.spain.barista.interaction.BaristaAutoCompleteTextViewInteractions.writeToAutoComplete;
 
 @RunWith(AndroidJUnit4.class)
 public class AutoCompleteTextViewTest {
