+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/ClearDatabaseRuleTest.java
@@ -5,7 +5,7 @@ import com.schibsted.spain.barista.rule.cleardata.ClearDatabaseRule;
 import org.junit.Rule;
 import org.junit.Test;
 
-import static com.schibsted.spain.barista.action.BaristaClickActions.clickOn;
+import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
 import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
 
 public class ClearDatabaseRuleTest {
