+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/DialogActivityTest.java
@@ -11,10 +11,10 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
-import static com.schibsted.spain.barista.action.BaristaClickActions.clickOn;
-import static com.schibsted.spain.barista.action.BaristaDialogActions.clickDialogNegativeButton;
-import static com.schibsted.spain.barista.action.BaristaDialogActions.clickDialogNeutralButton;
-import static com.schibsted.spain.barista.action.BaristaDialogActions.clickDialogPositiveButton;
+import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
+import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogNegativeButton;
+import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogNeutralButton;
+import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogPositiveButton;
 
 @RunWith(AndroidJUnit4.class)
 public class DialogActivityTest {
