+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/PermissionGranterTest.java
@@ -9,13 +9,13 @@ import android.support.test.InstrumentationRegistry;
 
 import com.schibsted.spain.barista.rule.BaristaRule;
 import com.schibsted.spain.barista.internal.failurehandler.BaristaException;
-import com.schibsted.spain.barista.action.PermissionGranter;
+import com.schibsted.spain.barista.interaction.PermissionGranter;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import static com.schibsted.spain.barista.action.BaristaClickActions.clickOn;
+import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
 import static org.junit.Assume.assumeFalse;
 import static org.junit.Assume.assumeTrue;
