+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/ScrollsTest.java
@@ -10,8 +10,8 @@ import org.junit.runner.RunWith;
 
 import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
 import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertNotDisplayed;
-import static com.schibsted.spain.barista.action.BaristaScrollActions.safelyScrollTo;
-import static com.schibsted.spain.barista.action.BaristaScrollActions.scrollTo;
+import static com.schibsted.spain.barista.interaction.BaristaScrollInteractions.safelyScrollTo;
+import static com.schibsted.spain.barista.interaction.BaristaScrollInteractions.scrollTo;
 
 @RunWith(AndroidJUnit4.class)
 public class ScrollsTest {
