+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/SleepTest.java
@@ -2,7 +2,7 @@ package com.schibsted.spain.barista.sample;
 
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
-import com.schibsted.spain.barista.action.BaristaSleepActions;
+import com.schibsted.spain.barista.interaction.BaristaSleepInteractions;
 import com.schibsted.spain.barista.sample.util.FailureHandlerValidatorRule;
 import java.util.concurrent.TimeUnit;
 import org.junit.Rule;
@@ -23,7 +23,7 @@ public class SleepTest {
   @Test
   public void sleepForASecondInMillis() {
     long start = System.currentTimeMillis();
-    BaristaSleepActions.sleep(1000);
+    BaristaSleepInteractions.sleep(1000);
     long end = System.currentTimeMillis();
 
     long duration = end - start;
@@ -35,7 +35,7 @@ public class SleepTest {
   @Test
   public void sleepForASecondInTimeUnits() {
     long start = System.currentTimeMillis();
-    BaristaSleepActions.sleep(1, TimeUnit.SECONDS);
+    BaristaSleepInteractions.sleep(1, TimeUnit.SECONDS);
     long end = System.currentTimeMillis();
 
     long duration = end - start;
