+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/introduction/IntroducingBaristaActions.java
@@ -6,28 +6,28 @@ import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static com.schibsted.spain.barista.action.BaristaAutoCompleteTextViewActions.writeToAutoComplete;
-import static com.schibsted.spain.barista.action.BaristaClickActions.clickBack;
-import static com.schibsted.spain.barista.action.BaristaClickActions.clickOn;
-import static com.schibsted.spain.barista.action.BaristaDialogActions.clickDialogNegativeButton;
-import static com.schibsted.spain.barista.action.BaristaDialogActions.clickDialogNeutralButton;
-import static com.schibsted.spain.barista.action.BaristaDialogActions.clickDialogPositiveButton;
-import static com.schibsted.spain.barista.action.BaristaEditTextActions.writeTo;
-import static com.schibsted.spain.barista.action.BaristaListActions.clickListItem;
-import static com.schibsted.spain.barista.action.BaristaListActions.clickListItemChild;
-import static com.schibsted.spain.barista.action.BaristaListActions.scrollListToPosition;
-import static com.schibsted.spain.barista.action.BaristaMenuClickActions.clickMenu;
-import static com.schibsted.spain.barista.action.BaristaDrawerActions.closeDrawer;
-import static com.schibsted.spain.barista.action.BaristaDrawerActions.openDrawer;
-import static com.schibsted.spain.barista.action.BaristaPickerActions.setDateOnPicker;
-import static com.schibsted.spain.barista.action.BaristaRadioButtonActions.clickRadioButtonItem;
-import static com.schibsted.spain.barista.action.BaristaRadioButtonActions.clickRadioButtonPosition;
-import static com.schibsted.spain.barista.action.BaristaScrollActions.scrollTo;
-import static com.schibsted.spain.barista.action.BaristaSleepActions.sleep;
-import static com.schibsted.spain.barista.action.BaristaSpinnerActions.clickSpinnerItem;
-import static com.schibsted.spain.barista.action.BaristaSwipeRefreshActions.refresh;
-import static com.schibsted.spain.barista.action.BaristaViewPagerActions.swipeViewPagerBack;
-import static com.schibsted.spain.barista.action.BaristaViewPagerActions.swipeViewPagerForward;
+import static com.schibsted.spain.barista.interaction.BaristaAutoCompleteTextViewInteractions.writeToAutoComplete;
+import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickBack;
+import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
+import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogNegativeButton;
+import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogNeutralButton;
+import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogPositiveButton;
+import static com.schibsted.spain.barista.interaction.BaristaEditTextInteractions.writeTo;
+import static com.schibsted.spain.barista.interaction.BaristaListInteractions.clickListItem;
+import static com.schibsted.spain.barista.interaction.BaristaListInteractions.clickListItemChild;
+import static com.schibsted.spain.barista.interaction.BaristaListInteractions.scrollListToPosition;
+import static com.schibsted.spain.barista.interaction.BaristaMenuClickInteractions.clickMenu;
+import static com.schibsted.spain.barista.interaction.BaristaDrawerInteractions.closeDrawer;
+import static com.schibsted.spain.barista.interaction.BaristaDrawerInteractions.openDrawer;
+import static com.schibsted.spain.barista.interaction.BaristaPickerInteractions.setDateOnPicker;
+import static com.schibsted.spain.barista.interaction.BaristaRadioButtonInteractions.clickRadioButtonItem;
+import static com.schibsted.spain.barista.interaction.BaristaRadioButtonInteractions.clickRadioButtonPosition;
+import static com.schibsted.spain.barista.interaction.BaristaScrollInteractions.scrollTo;
+import static com.schibsted.spain.barista.interaction.BaristaSleepInteractions.sleep;
+import static com.schibsted.spain.barista.interaction.BaristaSpinnerInteractions.clickSpinnerItem;
+import static com.schibsted.spain.barista.interaction.BaristaSwipeRefreshInteractions.refresh;
+import static com.schibsted.spain.barista.interaction.BaristaViewPagerInteractions.swipeViewPagerBack;
+import static com.schibsted.spain.barista.interaction.BaristaViewPagerInteractions.swipeViewPagerForward;
 import static java.util.concurrent.TimeUnit.SECONDS;
 
 @RunWith(AndroidJUnit4.class)
